Crate hotshot_testing_macros

Modules

keywords 🔒

macro specific custom keywords

Macros

cross_all_types

Generate a cartesian product of tests across all impls known to hotshot_types Arguments:

cross_tests

Generate a cartesian product of tests across all types Arguments:

Structs

CrossAllTypesSpec 🔒

we internally choose types

CrossAllTypesSpecBuilder 🔒

Builder for CrossAllTypesSpec.

CrossTestData 🔒

description of a crosstest

CrossTestDataBuilder 🔒

Builder for CrossTestData.

TestData 🔒

requisite data to generate a single test

TestDataBuilder 🔒

Builder for TestData.

Enums

CrossAllTypesSpecBuilderError 🔒

Error type for CrossAllTypesSpecBuilder

CrossTestDataBuilderError 🔒

Error type for CrossTestDataBuilder

SupportedConsensusTypes 🔒

Supported consensus types by macro

TestDataBuilderError 🔒

Error type for TestDataBuilder

Statics

_DECLS 🔒

Traits

ToLowerSnakeStr 🔒

trait make a string lower and snake case

Functions

cross_tests_internal 🔒