Macro hotshot_testing_macros::cross_tests

cross_tests!() { /* proc-macro */ }

Generate a cartesian product of tests across all types Arguments:

DemoType: [(ConsensusTypeName1, DemoStateType1), (ConsensusTypeName2, DemoStateType2) ...] - a list of tuples of cusonsensus state + State implementations

SignatureKey: [SignatureKey1, SignatureKey2, ...] - a list of SignatureKey implementations

CommChannel: [CommChannel1, CommChannel2, ..] - a list of CommunicationChannel implementations

Time: [ Time1, Time2, ...] - a list of ConsensusTime implementations]`

TestName: example_test - the name of the test

TestBuilder: { some_test_builder_expression } - the TestBuilder to use

Storage: Storage1, Storage2, ... - a list of Storage implementations to use

Slow: whether or not this set of tests are hidden behind the slow feature flag Example usage:

hotshot_testing_macros::cross_tests!(
    DemoType: [(ValidatingConsensus, hotshot::demos::vdemo::VDemoState) ],
    SignatureKey: [ hotshot_types::traits::signature_key::ed25519::Ed25519Pub ],
    CommChannel: [ hotshot::traits::implementations::MemoryCommChannel ],
    Storage: [ hotshot::traits::implementations::MemoryStorage ],
    Time: [ hotshot_types::data::ViewNumber ],
    TestName: ten_tx_seven_nodes_fast,
    TestBuilder: hotshot_testing::test_builder::TestBuilder {
        total_nodes: 7,
        start_nodes: 7,
        num_succeeds: 10,
        txn_ids: either::Either::Right(1),
        ..hotshot_testing::test_builder::TestBuilder::default()
    },
    Slow: false,
);